create dedicated storage util
