+++ ml_peg/app/build_app.py
@@ -17,10 +17,9 @@ from dash import (
 from dash.dash_table import DataTable
 from dash.dcc import Dropdown, Interval, Link, Loading, Location, Store
 from dash.exceptions import PreventUpdate
-from dash.html import H1, H3, A, Br, Button, Details, Div, Img, Span, Summary
+from dash.html import H1, H3, A, Br, Details, Div, Img, Span, Summary
 from yaml import safe_load
 
-from ml_peg import __version__
 from ml_peg.analysis.utils.utils import calc_table_scores, get_table_style
 from ml_peg.app import APP_ROOT
 from ml_peg.app.filters import (
@@ -39,13 +38,16 @@ from ml_peg.app.utils.build_components import (
 )
 from ml_peg.app.utils.onboarding import (
     build_onboarding_modal,
-    build_tutorial_button,
     register_onboarding_callbacks,
 )
 from ml_peg.app.utils.register_callbacks import (
     register_benchmark_to_category_callback,
     register_filter_loading_callback,
     register_filter_tables_callback,
+)
+from ml_peg.app.utils.storage import (
+    build_header_controls,
+    register_storage_callbacks,
 )
 from ml_peg.app.utils.utils import (
     build_level_of_theory_warnings,
@@ -1084,47 +1086,7 @@ def build_nav(
         ),
         Interval(id="startup-mask-poll", interval=250, n_intervals=0),
         build_onboarding_modal(),
-        # Fixed header controls (top-right): clear-cache button next to the
-        # tutorial button. The hidden Divs are the clear-storage and
-        # version-check clientside-callback targets.
-        Div(
-            [
-                Button(
-                    "Clear cache",
-                    id="clear-storage-button",
-                    n_clicks=0,
-                    title=(
-                        "Clear browser-stored app state (weights, thresholds, "
-                        "tutorial progress) and reload. Use after an update if "
-                        "the app shows stale data."
-                    ),
-                    style={
-                        "padding": "8px 16px",
-                        "borderRadius": "6px",
-                        "border": "1px solid #cbd5e1",
-                        "background": "white",
-                        "color": "#475569",
-                        "cursor": "pointer",
-                        "fontWeight": 600,
-                        "fontSize": "14px",
-                        "boxShadow": "0 2px 8px rgba(0, 0, 0, 0.1)",
-                        "transition": "all 0.2s ease",
-                    },
-                ),
-                build_tutorial_button(),
-                Div(id="clear-storage-dummy", style={"display": "none"}),
-                Div(id="storage-version-dummy", style={"display": "none"}),
-            ],
-            style={
-                "position": "fixed",
-                "top": "20px",
-                "right": "20px",
-                "display": "flex",
-                "alignItems": "center",
-                "gap": "10px",
-                "zIndex": "1600",  # Above loading overlays (1200/1400).
-            },
-        ),
+        build_header_controls(),
         Location(id="app-location", refresh=False),
         Store(
             id="summary-table-scores-store",
@@ -1258,51 +1220,7 @@ def build_nav(
         Input("startup-mask-poll", "n_intervals"),
     )
 
-    # Clear all browser-persisted dcc.Store data (session + local) and reload, so
-    # stale cached state after an update can be wiped from the footer button.
-    clientside_callback(
-        """
-        function (n_clicks) {
-            if (n_clicks && window.confirm(
-                "Clear cached app data and reload? Saved weights and thresholds"
-                + " will be reset."
-            )) {
-                window.localStorage.clear();
-                window.sessionStorage.clear();
-                window.location.reload();
-            }
-            return "";
-        }
-        """,
-        Output("clear-storage-dummy", "children"),
-        Input("clear-storage-button", "n_clicks"),
-        prevent_initial_call=True,
-    )
-
-    # Auto-clear browser-persisted stores when the released version changes, so a
-    # new release drops stale cached state automatically. The
-    # version is recorded in localStorage; a real change (not a first visit) clears
-    # both storages and reloads so the dcc.Stores reinitialise from fresh defaults.
-    clientside_callback(
-        f"""
-        function (pathname) {{
-            const current = "{__version__}";
-            const stored = window.localStorage.getItem("ml-peg-store-version");
-            if (stored !== current) {{
-                window.localStorage.clear();
-                window.sessionStorage.clear();
-                window.localStorage.setItem("ml-peg-store-version", current);
-                if (stored !== null) {{
-                    window.location.reload();
-                }}
-            }}
-            return "";
-        }}
-        """,
-        Output("storage-version-dummy", "children"),
-        Input("app-location", "pathname"),
-        prevent_initial_call=False,
-    )
+    register_storage_callbacks()
 
     @callback(
         Output("model-filter-checklist", "value"),
